dispatch ppu timing related functions when m_cycles are spent

+++ src/common.h
@@ -212,10 +212,10 @@ typedef struct io_regs {
 
   // This is technically the system clock which is essentially the full 16 bit version of DIV, the DIV register just
   // returns the most significant 8 bits.
-  uint16_t div;  // divider register
-  uint8_t  tima; // timer counter
-  uint8_t  tma;  // timer modulo
-  uint8_t  tac;  // timer control
+  uint8_t div;  // divider register
+  uint8_t tima; // timer counter
+  uint8_t tma;  // timer modulo
+  uint8_t tac;  // timer control
   // Sound
   uint8_t nr10;
   uint8_t nr11;
