+++ src/timing.h
@@ -5,6 +5,6 @@
 
 typedef struct gb_state gb_state_t;
 
-void gb_spend_mcycles(gb_state_t *gb_state, uint64_t n);
+void gb_spend_mcycles(gb_state_t *gb_state, uint16_t n);
 
 #endif // GB_TIMING_H
